requirements.txt
openai
requests
beautifulsoup4
pyyaml
ddgs
pydantic
pydantic-ai
pytest
ruff
mypy
logfire
supabase
numpy
tenacity
networkx
click
rich
streamlit
fastapi
uvicorn
httpx
asyncio
redis
celery
pandas
plotly
python-dotenv
websockets
aiofiles
prometheus-client
psycopg2-binary
sqlalchemy
alembic